moved distance metric to coordinates module
