+++ dynhalo/finder/coordinates.py
@@ -81,5 +81,37 @@ def get_vr_vt_from_coordinates(
     return vr, vt, v2
 
 
+def distance_metric(
+    rel_pos: np.ndarray,
+    rel_vel: np.ndarray,
+    sigma_x_sq: float,
+    sigma_v_sq: float,
+    lamb: float = 1.0,
+) -> float:
+    """Computes the distance between particles and halo centre in
+    six-dimensional phase space.
+
+    Parameters
+    ----------
+    rel_pos : np.ndarray
+        Relative position of paticles to halo
+    rel_vel : np.ndarray
+        Relative velocity of paticles to halo
+    sigma_x_sq : float
+        Position dispersion parameter 
+    sigma_v_sq : float
+        Velocity dispersion parameter 
+    lamb : float, optional
+        Scaling parameter in distance metric, by default 1.0
+
+    Returns
+    -------
+    float
+        Six-dimensional distance
+    """
+    return np.sum(np.square(rel_pos), axis=1) / sigma_x_sq + \
+            lamb * np.sum(np.square(rel_vel), axis=1) / sigma_v_sq
+
+
 if __name__ == "__main__":
     pass
